Update: Added exectimer decorator to fsutils/utils/__init__.py
This commit introduces a new decorator, exectimer, to the fsutils/utils/__init__.py module.
The exectimer decorator is used to measure the execution time of a function or method.
This decorator can be applied to any function or method to measure its execution time and log the results.

NOTE:
Intended for development purposes